+++ btu_meter/async_btu_main.py
@@ -1,371 +1,4 @@
 import machine
-import ujson
-import network
-import utime
-import urequests
-import ubinascii
-import os
-
-PRIMARY_SCADA_IP = "192.168.2.200"
-
-# Remove existing files
-if 'boot.py' in os.listdir():
-    os.remove('boot.py')
-if 'app_config.json' in os.listdir():
-    os.remove('app_config.json')
-if 'comms_config.json' in os.listdir():
-    os.remove('comms_config.json')
-if 'main.py' in os.listdir():
-    os.remove('main.py')
-if 'main_previous.py' in os.listdir():
-    os.remove('main_previous.py')
-
-# *************************
-# 1/3 - MAIN.PY PROVISION
-# *************************
-
-def write_tank_module_3_main():
-    main_code = """
-
-import machine
-import utime
-import network
-import ujson
-import urequests
-import ubinascii
-import utime
-import gc
-
-# ---------------------------------
-# Constants
-# ---------------------------------
-
-# Configuration files
-COMMS_CONFIG_FILE = "comms_config.json"
-APP_CONFIG_FILE = "app_config.json"
-
-# Default parameters
-DEFAULT_ACTOR_NAME = "tank"
-DEFAULT_ASYNC_CAPTURE_DELTA_MICRO_VOLTS = 500
-DEFAULT_CAPTURE_PERIOD_S = 60
-DEFAULT_SAMPLES = 1000
-DEFAULT_NUM_SAMPLE_AVERAGES = 10
-
-# Other constants
-ADC0_PIN_NUMBER = 26
-ADC1_PIN_NUMBER = 27
-ADC2_PIN_NUMBER = 28
-
-# ---------------------------------
-# Main class
-# ---------------------------------
-
-class TankModule3:
-
-    def __init__(self):
-        # Unique ID
-        pico_unique_id = ubinascii.hexlify(machine.unique_id()).decode()[-6:]
-        self.hw_uid = f"pico_{pico_unique_id}"
-        # Pins
-        self.adc0 = machine.ADC(ADC0_PIN_NUMBER)
-        self.adc1 = machine.ADC(ADC1_PIN_NUMBER)
-        self.adc2 = machine.ADC(ADC2_PIN_NUMBER)
-        # Load configuration files
-        self.load_comms_config()
-        self.load_app_config()
-        # Measuring and repoting voltages
-        self.prev_mv0 = -1
-        self.prev_mv1 = -1
-        self.prev_mv2 = -1
-        self.mv0 = None
-        self.mv1 = None
-        self.mv2 = None
-        self.node_names = []
-        self.microvolts_posted_time = utime.time()
-        # Synchronous reporting on the minute
-        self.capture_offset_seconds = 0
-        self.sync_report_timer = machine.Timer(-1)
-
-    def set_names(self):
-        if self.actor_node_name is None:
-            raise Exception("Needs actor node name or pico number to run. Reboot!")
-        self.node_names = [
-            f"{self.actor_node_name}-depth1", 
-            f"{self.actor_node_name}-depth2",
-            f"{self.actor_node_name}-depth3"
-        ]
-
-    # ---------------------------------
-    # Communication
-    # ---------------------------------
-                                                                 
-    def load_comms_config(self):
-        '''Load the communication configuration file (WiFi/Ethernet and API base URL)'''
-        try:
-            with open(COMMS_CONFIG_FILE, "r") as f:
-                comms_config = ujson.load(f)
-        except (OSError, ValueError) as e:
-            raise RuntimeError(f"Error loading comms_config file: {e}")
-        self.wifi_or_ethernet = comms_config.get("WifiOrEthernet", 'wifi')
-        self.wifi_name = comms_config.get("WifiName", None)
-        self.wifi_password = comms_config.get("WifiPassword", None)
-        self.base_url = comms_config.get("BaseUrl")
-        if self.wifi_or_ethernet=='wifi':
-            if self.wifi_name is None:
-                raise KeyError("WifiName not found in comms_config.json")
-            if self.wifi_password is None:
-                raise KeyError("WifiPassword not found in comms_config.json")
-        elif self.wifi_or_ethernet=='ethernet':
-            pass
-        else:
-            raise KeyError("WifiOrEthernet must be either 'wifi' or 'ethernet' in comms_config.json")
-        if self.base_url is None:
-            raise KeyError("BaseUrl not found in comms_config.json")
-        
-    def connect_to_wifi(self):
-        wlan = network.WLAN(network.STA_IF)
-        wlan.active(True)
-        if not wlan.isconnected():
-            print("Connecting to wifi...")
-            wlan.connect(self.wifi_name, self.wifi_password)
-            while not wlan.isconnected():
-                utime.sleep_ms(500)
-        print(f"Connected to wifi {self.wifi_name}")
-
-    def connect_to_ethernet(self):
-        nic = network.WIZNET5K()
-        for attempt in range(3):
-            try:
-                nic.active(True)
-                break
-            except Exception as e:
-                print(f"Retrying NIC activation due to: {e}")
-                utime.sleep(0.5)
-        if not nic.isconnected():
-            print("Connecting to Ethernet...")
-            nic.ifconfig('dhcp')
-            timeout = 10
-            start = utime.time()
-            while not nic.isconnected():
-                if utime.time() - start > timeout:
-                    raise RuntimeError("Failed to connect to Ethernet (timeout)")
-                utime.sleep(0.5)
-        print("Connected to Ethernet")
-
-    # ---------------------------------
-    # Parameters
-    # ---------------------------------
-    
-    def load_app_config(self):
-        '''
-        Set parameters to their value in the app_config file if it is specified
-        Otherwise set them to their default value
-        '''
-        try:
-            with open(APP_CONFIG_FILE, "r") as f:
-                app_config = ujson.load(f)
-        except:
-            app_config = {}
-        self.actor_node_name = app_config.get("ActorNodeName", DEFAULT_ACTOR_NAME)
-        self.async_capture_delta_micro_volts = app_config.get("AsyncCaptureDeltaMicroVolts", DEFAULT_ASYNC_CAPTURE_DELTA_MICRO_VOLTS)
-        self.capture_period_s = app_config.get("CapturePeriodS", DEFAULT_CAPTURE_PERIOD_S)
-        self.samples = app_config.get("Samples", DEFAULT_SAMPLES)
-        self.num_sample_averages = app_config.get("NumSampleAverages", DEFAULT_NUM_SAMPLE_AVERAGES)
-
-    def save_app_config(self):
-        config = {
-            "ActorNodeName": self.actor_node_name,
-            "CapturePeriodS": self.capture_period_s,
-            "Samples": self.samples,
-            "NumSampleAverages":self.num_sample_averages,
-            "AsyncCaptureDeltaMicroVolts": self.async_capture_delta_micro_volts,
-        }
-        with open(APP_CONFIG_FILE, "w") as f:
-            ujson.dump(config, f)
-    
-    def update_app_config(self):
-        url = self.base_url + f"/{self.actor_node_name}/tank-module-params"
-        payload = {
-            "HwUid": self.hw_uid,
-            "ActorNodeName": self.actor_node_name,
-            "CapturePeriodS": self.capture_period_s,
-            "Samples": self.samples,
-            "NumSampleAverages": self.num_sample_averages,
-            "AsyncCaptureDeltaMicroVolts": self.async_capture_delta_micro_volts,
-            "TypeName": "tank.module.params",
-            "Version": "110"
-        }
-        headers = {"Content-Type": "application/json"}
-        json_payload = ujson.dumps(payload)
-        try:
-            response = urequests.post(url, data=json_payload, headers=headers)
-            if response.status_code == 200:
-                updated_config = response.json()
-                self.actor_node_name = updated_config.get("ActorNodeName", self.actor_node_name)
-                self.capture_period_s = updated_config.get("CapturePeriodS", self.capture_period_s)
-                self.samples = updated_config.get("Samples", self.samples)
-                self.num_sample_averages = updated_config.get("NumSampleAverages", self.num_sample_averages)
-                self.async_capture_delta_micro_volts = updated_config.get("AsyncCaptureDeltaMicroVolts", self.async_capture_delta_micro_volts)
-                self.capture_offset_seconds = updated_config.get("CaptureOffsetS", 0)
-                self.save_app_config()
-            response.close()
-        except Exception as e:
-            print(f"Error sending tank module params: {e}")
-
-    # ---------------------------------
-    # Code updates
-    # ---------------------------------
-
-    def update_code(self):
-        url = self.base_url + f"/{self.actor_node_name}/code-update"
-        payload = {
-            "HwUid": self.hw_uid,
-            "ActorNodeName": self.actor_node_name,
-            "TypeName": "new.code",
-            "Version": "100"
-        }
-        json_payload = ujson.dumps(payload)
-        headers = {"Content-Type": "application/json"}
-        response = urequests.post(url, data=json_payload, headers=headers)
-        if response.status_code == 200:
-            # If there is a pending code update then the response is a python file, otherwise json
-            try:
-                ujson.loads(response.content.decode('utf-8'))
-            except:
-                python_code = response.content
-                with open('main_update.py', 'wb') as file:
-                    file.write(python_code)
-                machine.reset()
-
-    # ---------------------------------
-    # Measuring microvolts
-    # ---------------------------------
-
-    def adc0_micros(self):
-        sample_averages = []
-        for _ in range(self.num_sample_averages):
-            readings = []
-            for _ in range(self.samples):
-                # Read the raw ADC value (0-65535)
-                readings.append(self.adc0.read_u16())
-            voltages = list(map(lambda x: x * 3.3 / 65535, readings))
-            mean_1000 = int(10**6 * sum(voltages) / self.samples)
-            sample_averages.append(mean_1000)
-        return int(sum(sample_averages)/self.num_sample_averages)
-    
-    def adc1_micros(self):
-        sample_averages = []
-        for _ in range(self.num_sample_averages):
-            readings = []
-            for _ in range(self.samples):
-                # Read the raw ADC value (0-65535)
-                readings.append(self.adc1.read_u16())
-            voltages = list(map(lambda x: x * 3.3 / 65535, readings))
-            mean_1000 = int(10**6 * sum(voltages) / self.samples)
-            sample_averages.append(mean_1000)
-        return int(sum(sample_averages)/self.num_sample_averages)
-    
-    def adc2_micros(self):
-        sample_averages = []
-        for _ in range(self.num_sample_averages):
-            readings = []
-            for _ in range(self.samples):
-                # Read the raw ADC value (0-65535)
-                readings.append(self.adc2.read_u16())
-            voltages = list(map(lambda x: x * 3.3 / 65535, readings))
-            mean_1000 = int(10**6 * sum(voltages) / self.samples)
-            sample_averages.append(mean_1000)
-        return int(sum(sample_averages)/self.num_sample_averages)  
-    
-    # ---------------------------------
-    # Posting microvolts
-    # ---------------------------------
-
-    def post_microvolts(self, idx=3):
-        url = self.base_url + f"/{self.actor_node_name}/microvolts"
-        if idx==0:
-            mv_list = [self.mv0]
-        elif idx==1:
-            mv_list = [self.mv1]
-        elif idx==2:
-            mv_list = [self.mv2]
-        else:
-            mv_list = [self.mv0, self.mv1, self.mv2]
-        payload = {
-            "HwUid": self.hw_uid,
-            "AboutNodeNameList": [self.node_names[idx]] if idx<=2 else self.node_names,
-            "MicroVoltsList": mv_list, 
-            "TypeName": "microvolts", 
-            "Version": "100"
-        }
-        headers = {'Content-Type': 'application/json'}
-        json_payload = ujson.dumps(payload)
-        try:
-            response = urequests.post(url, data=json_payload, headers=headers)
-            response.close()
-        except Exception as e:
-            print(f"Error posting microvolts: {e}")
-        gc.collect()
-        self.microvolts_posted_time = utime.time()
-        
-    def sync_report(self, timer):
-        self.post_microvolts()
-
-    def start_sync_report_timer(self):
-        '''Initialize the timer to call self.keep_alive periodically'''
-        self.sync_report_timer.init(
-            period=self.capture_period_s * 1000, 
-            mode=machine.Timer.PERIODIC,
-            callback=self.sync_report
-        )
-
-    def main_loop(self):
-        self.mv0 = self.adc0_micros()
-        self.mv1 = self.adc1_micros()
-        self.mv2 = self.adc2_micros()
-        while True:
-            self.mv0 = self.adc0_micros()
-            self.mv1 = self.adc1_micros()
-            self.mv2 = self.adc2_micros()
-            if abs(self.mv0 - self.prev_mv0) > self.async_capture_delta_micro_volts:
-                self.post_microvolts(idx=0)
-                self.prev_mv0 = self.mv0
-            if abs(self.mv1 - self.prev_mv1) > self.async_capture_delta_micro_volts:
-                self.post_microvolts(idx=1)
-                self.prev_mv1 = self.mv1
-            if abs(self.mv2 - self.prev_mv2) > self.async_capture_delta_micro_volts:
-                self.post_microvolts(idx=2)
-                self.prev_mv2 = self.mv2
-            utime.sleep_ms(100)
-
-    def start(self):
-        if self.wifi_or_ethernet=='wifi':
-            self.connect_to_wifi()
-        elif self.wifi_or_ethernet=='ethernet':
-            self.connect_to_ethernet()
-        self.update_code()
-        self.update_app_config()
-        self.set_names()
-        self.mv0 = self.adc0_micros()
-        self.mv1 = self.adc1_micros()
-        self.mv2 = self.adc2_micros()
-        self.post_microvolts()
-        utime.sleep(self.capture_offset_seconds)
-        self.start_sync_report_timer()
-        self.main_loop()
-
-if __name__ == "__main__":
-    t = TankModule3()
-    t.start()
-    
-
-    """
-    with open('main.py', 'w') as file:
-        file.write(main_code)
-
-def write_btu_meter_main():
-    main_code = """
-import machine
 import utime
 import math
 import network
@@ -1250,234 +883,3 @@ class AsyncBtuMeter:
 if __name__ == "__main__":
     b = AsyncBtuMeter()
     b.start()
-
-    """
-    with open('main.py', 'w') as file:
-        file.write(main_code)
-
-
-# *************************
-# 2/3 - APP_CONFIG PROVISION
-# *************************
-
-# -------------------------
-# Tank module
-# -------------------------
-
-def provision_tank_module():
-    """Configure tank module app_config.json"""
-    # Get tank name
-    while True:
-        tank_name = input("Tank Name: 'buffer', 'tank1', 'tank2', 'tank3': ")
-        if tank_name in {'buffer', 'tank1', 'tank2', 'tank3'}:
-            break
-        print("Invalid tank name")
-
-    config = {
-            "ActorNodeName": tank_name,
-        }
-
-    # Save config
-    with open("app_config.json", "w") as f:
-        ujson.dump(config, f)
-
-    return tank_name
-
-
-# -------------------------
-# BTU meter
-# -------------------------
-
-def provision_btu_meter():
-    """Configure BTU meter app_config"""
-    while True:
-        btu_name = input("BTU Name: 'dist-btu', 'store-btu', 'primary-btu', 'sieg-btu': ")
-        if btu_name in {'primary-btu', 'store-btu', 'dist-btu', 'sieg-btu'}:
-            break
-        print("Invalid BTU name")
-
-    config = {
-        "ActorNodeName": btu_name,
-    }
-
-    # Save config
-    with open("app_config.json", "w") as f:
-        ujson.dump(config, f)
-
-    return btu_name
-
-
-# *************************
-# 3/3 - MAIN CODE
-# *************************
-
-if __name__ == "__main__":
-
-    # Get hardware ID
-    pico_unique_id = ubinascii.hexlify(machine.unique_id()).decode()[-6:]
-    hw_uid = f"pico_{pico_unique_id}"
-    print(f"\nThis Pico's unique hardware ID is {hw_uid}.")
-
-    # -------------------------
-    # Write boot.py
-    # -------------------------
-
-    bootpy_code = """import os
-
-if 'main_update.py' in os.listdir():
-    
-    if 'main_previous.py' in os.listdir():
-        os.remove('main_previous.py')
-
-    if 'main.py' in os.listdir():
-        os.rename('main.py', 'main_previous.py')
-
-    os.rename('main_update.py', 'main.py')
-
-elif 'main_revert.py' in os.listdir():
-
-    if 'main.py' in os.listdir():
-        os.remove('main.py')
-
-    os.rename('main_revert.py', 'main.py')
-    """
-
-    with open('boot.py', 'w') as file:
-        file.write(bootpy_code)
-    print(f"Wrote 'boot.py' on the Pico.")
-    
-    print(f"\n{'-'*40}\n[1/4] Success! Found hardware ID and wrote 'boot.py'.\n{'-'*40}\n")
-
-    # -------------------------
-    # Write comms_config.json
-    # -------------------------
-
-    have_wifi_or_ethernet = False
-    while not have_wifi_or_ethernet:
-        wifi_or_ethernet = input("Does this Pico use WiFi (enter 'w') or Ethernet (enter 'e'): ")
-        if wifi_or_ethernet not in {'w','e'}:
-            print("Invalid entry. Please enter either 'w' or 'e'.")
-        else:
-            have_wifi_or_ethernet = True
-    
-    # Connect to wifi
-    if wifi_or_ethernet == 'w':
-        wlan = network.WLAN(network.STA_IF)
-        wlan.active(True)
-        wlan.disconnect()
-        while wlan.isconnected():
-            utime.sleep(0.1)
-        
-        while not wlan.isconnected():
-            wifi_name = input("Enter wifi name (leave blank for 'GridWorks'): ")
-            if wifi_name == "":
-                wifi_name = "GridWorks"
-            wifi_pass = input("Enter wifi password: ")
-            time_waiting_connection = 0
-            wlan.connect(wifi_name, wifi_pass)
-            while not wlan.isconnected():
-                if time_waiting_connection>0 and time_waiting_connection%2==0:
-                    print(f"Trying to connect ({int(time_waiting_connection/2)}/5)...")
-                utime.sleep(0.5)
-                time_waiting_connection += 0.5
-                if time_waiting_connection > 10:
-                    print("Failed to connect to wifi, please try again.\n")
-                    break
-        print(f"Connected to wifi '{wifi_name}'.\n")
-
-    # Connect to ethernet
-    elif wifi_or_ethernet == 'e':
-        nic = network.WIZNET5K()
-        for attempt in range(3):
-            try:
-                nic.active(True)
-                break
-            except Exception as e:
-                print(f"Retrying NIC activation due to: {e}")
-                utime.sleep(0.5)
-        if not nic.isconnected():
-            print("Connecting to Ethernet...")
-            nic.ifconfig('dhcp')
-            timeout = 10
-            start = utime.time()
-            while not nic.isconnected():
-                if utime.time() - start > timeout:
-                    raise RuntimeError("Failed to connect to Ethernet (timeout)")
-                utime.sleep(0.5)
-        print("Connected to Ethernet")
-
-    # Connect to API
-
-    connected_to_api = False
-    while not connected_to_api:
-
-        ip_address = input("Enter IP address (return for default): ").strip()
-        if ip_address == '':
-            ip_address = PRIMARY_SCADA_IP
-
-        base_url = f"http://{ip_address}:8000"
-
-        url = base_url + "/new-pico"
-        payload = {
-            "HwUid": hw_uid,
-            "TypeName": "new.pico",
-            "Version": "100"
-        }
-        headers = {"Content-Type": "application/json"}
-        json_payload = ujson.dumps(payload)
-        try:
-            response = urequests.post(url, data=json_payload, headers=headers)
-            if response.status_code == 200:
-                connected_to_api = True
-            else:
-                print(f"Connected to the API, but it returned a status code {response.status_code}, indicating an issue.")
-            response.close()
-        except Exception as e:
-            print(f"There was an error connecting to the API: {e}. Please check the hostname and try again.")
-
-    print(f"Connected to the API hosted at '{base_url}'.")
-    hostname = input("Enter hostname for backup (e.g., 'beech'): ").strip()
-    backup_url = f"http://{hostname}.local:8000"
-
-    # Write the parameters to comms_config.json
-    if wifi_or_ethernet=='w':
-        comms_config_content = {
-            "WifiOrEthernet": 'wifi',
-            "WifiName": wifi_name,
-            "WifiPassword": wifi_pass, 
-            "BaseUrl": f"http://{PRIMARY_SCADA_IP}:8000",
-            "BackupUrl": backup_url
-        }
-    elif wifi_or_ethernet=='e':
-        comms_config_content = {
-            "WifiOrEthernet": 'ethernet',
-            "BaseUrl": f"http://{PRIMARY_SCADA_IP}:8000",
-            "BackupUrl": backup_url
-        }
-    with open('comms_config.json', 'w') as file:
-        ujson.dump(comms_config_content, file)
-
-    print(f"\n{'-'*40}\n[2/4] Success! Wrote 'comms_config.json' on the Pico.\n{'-'*40}\n")
-
-    # -------------------------
-    # Write app_config.json and main code
-    # -------------------------
-    while True:
-        device_type = input("Is this Pico associated to a TankModule3 (enter '0') or an AsyncBtuMeter (enter '1'): ")
-        if device_type in {'0', '1'}:
-            break
-        print('Please enter 0 or 1.')
-
-    if device_type == '0':
-        actor_name = provision_tank_module()
-        print("This is a tank module")
-        write_tank_module_3_main()
-    elif device_type == '1':
-        actor_name = provision_btu_meter()
-        print("This is a BTU meter.")
-        write_btu_meter_main()
-        
-
-    print(f"\n{'-'*40}\n[4/4] Success! Wrote 'main.py' on the Pico.\n{'-'*40}\n")
-
-    print("The Pico is set up. It is now ready to use.")
